Add top module `generate_md_system`
Having `two_phase` as a single, loose module installed with the package
isn't great. This creates a proper top level module with the same name
as the package and moves `two_phase` inside of it as a submodule.

Besides being better from a name space perspective this allows us to
create more submodules as needed.